Backend/requirements.txt
django
djangorestframework
pycryptodome
django-cors-headers
